requirements.txt
Django
django_allauth
django-cors-headers
dj_rest_auth
djangorestframework
djangorestframework-simplejwt
requests